Add Locales tests
