+++ tests/integration/WC_Facebookcommerce_MessengerChat_Test.php
@@ -27,33 +27,5 @@ class WC_Facebookcommerce_MessengerChat_Test extends \Codeception\TestCase\WPTes
 	}
 
 
-	/** Test methods **************************************************************************************************/
-
-
-	/** @see \WC_Facebookcommerce_MessengerChat::get_supported_locales() */
-	public function test_get_supported_locales() {
-
-		$locales = \WC_Facebookcommerce_MessengerChat::get_supported_locales();
-
-		$this->assertIsArray( $locales );
-		$this->assertArrayHasKey( 'en_US', $locales );
-	}
-
-
-	/** @see \WC_Facebookcommerce_MessengerChat::get_supported_locales() */
-	public function test_get_supported_locales_filter() {
-
-		add_filter( 'wc_facebook_messenger_supported_locales', static function( $locales ) {
-
-			$locales['custom'] = 'Custom';
-
-			return $locales;
-
-		} );
-
-		$this->assertArrayHasKey( 'custom', \WC_Facebookcommerce_MessengerChat::get_supported_locales() );
-	}
-
-
 }
